added mode selection

+++ ArcAim/src/Button.cpp
@@ -12,8 +12,8 @@ Button::Button(sf::Color btnColor, float width, float height, float x, float y, 
 	m_text.setString(string);
 	m_text.setFillColor(txtColor);
 	m_text.setCharacterSize(25);
-	m_text.setPosition((m_shape.getPosition().x + m_shape.getGlobalBounds().width / 2) - (m_text.getGlobalBounds().width / 2.f),
-		(m_shape.getPosition().y + m_shape.getGlobalBounds().height / 2) - (m_text.getGlobalBounds().height / 1.3f));
+	m_text.setPosition((int)((m_shape.getPosition().x + m_shape.getGlobalBounds().width / 2.f) - (m_text.getGlobalBounds().width / 2.f)),
+		(int)((m_shape.getPosition().y + m_shape.getGlobalBounds().height / 2.f) - (m_text.getGlobalBounds().height / 1.3f)));
 }
 
 Button::~Button()
